requirements.txt
anyio==3.7.0
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.5
django-allauth==0.61.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-paypal==2.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-widget-tweaks==1.5.0
filelock==3.13.1
geographiclib==2.0
geopy==2.4.1
ghp-import==2.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.3
oauthlib==3.2.2
packaging==23.1
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.25.0
watchdog==3.0.0